mbean fix

+++ src/main/java/com/example/wp/mbeans/PercentageCounter.java
@@ -4,10 +4,11 @@ import org.springframework.jmx.export.annotation.ManagedResource;
 import org.springframework.stereotype.Component;
 
 import javax.management.NotificationBroadcasterSupport;
+
 @Component
 @ManagedResource
 public class PercentageCounter extends NotificationBroadcasterSupport implements PercentageCounterMBean {
-    private long percentageCounter;
+    private double percentageCounter;
 
     @Override
     public void updateCounters(long successful, long all) {
@@ -15,7 +16,7 @@ public class PercentageCounter extends NotificationBroadcasterSupport implements
     }
 
     @Override
-    public long getPercentageCounter() {
+    public double getPercentageCounter() {
         return percentageCounter;
     }
 }
