+++ src/main/java/com/example/wp/mbeans/PercentageCounterMBean.java
@@ -3,5 +3,5 @@ package com.example.wp.mbeans;
 public interface PercentageCounterMBean {
     void updateCounters(long successful, long all);
 
-    long getPercentageCounter();
+    double getPercentageCounter();
 }
